Update C-wrapped msg warning
Before the msg C++ read() interface checked if a msg was linked to.  However, this didn't make sense for
stand-along messages.  Rather, the issue is reading a msg if it was never written to.  This can happen
by another msg object being redirected to write to it, or it could be written to from python.  Now the
error checking is if the isWritten flag is true of the headerPointer information, not the head.
This ensure that this works with re-directed messages as well.