Remove msg payload and replace with classic elements type
The modules were using the ClassicElementsMsgPayload which is typedef'd
as classicElements. This typedef and its inclusion has resulted in polluting
the system with a plurality of orbital element types and has misused a
message type. To support an aspect of decoupling between system
components (at runtime and compile time) message types should never be
included for usage beyond reading and writing messages.