Remove overwriteSwig.py
As part of the build, this script attempts to insert an extra
conditional at certain points in three specific SWIG output bindings.
However, inspecting those files after-the-fact reveals no such extra
conditional, nor the expected stimulus for that conditional to be
inserted. As a result, it is clear that this script is a complete and
utter no-op.